Implement MetadataManager for create subcommand

+++ runt/src/container/mod.rs
@@ -1,17 +1,21 @@
+use std::fs::{self, File};
+use std::os::unix::io::AsRawFd;
 use std::path::PathBuf;
 
 use anyhow::Result;
+use nix::fcntl::{flock, FlockArg};
+use serde_derive::{Deserialize, Serialize};
 
-use crate::container::specs::{Spec, State, Status, OCI_VERSION};
+use crate::container::specs::{Spec, Status};
 
 pub mod specs;
 
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Container {
     pub id: String,
     pub spec: Spec,
     pub bundle: PathBuf,
-    // pub state: State,
-    // pub metadata: Metadata
+    pub status: Status,
 }
 
 impl Container {
@@ -20,21 +24,68 @@ impl Container {
             id: id.into(),
             bundle: bundle.to_path_buf(),
             spec,
+            status: Status::Creating,
         }
     }
 
-    pub fn create(&self) -> Result<State> {
-        Ok(State {
-            oci_version: OCI_VERSION.into(),
-            id: self.id.clone(),
+    pub fn create(&mut self) -> Result<()> {
+        // assert_eq!(container.status, Status::Creating)
+        self.save_metadata(&self)?;
+
+        // -----
+        // container creating
+        // -----
+
+        self.status = Status::Created;
+        self.save_metadata(&self)?;
+        Ok(())
+    }
+}
+
+pub static DEFAULT_META_ROOT: &str = "/tmp/runt";
+
+trait MetadataManager {
+    fn save_metadata(&self, container: &Container) -> Result<()>;
+    fn load() -> Result<Container>;
+    fn lock(&self, file: &File) -> Result<()>;
+    fn unlock(&self, file: &File) -> Result<()>;
+}
+
+impl MetadataManager for Container {
+    fn save_metadata(&self, container: &Container) -> Result<()> {
+        let metadata_dir = PathBuf::from(DEFAULT_META_ROOT).join(&container.id);
+        if !metadata_dir.exists() {
+            fs::create_dir_all(&metadata_dir)?;
+        }
+        let statefile = File::create(metadata_dir.join("state.json"))?;
+
+        self.lock(&statefile)?;
+
+        serde_json::to_writer(&statefile, &container)?;
+
+        self.unlock(&statefile)?;
+        Ok(())
+    }
+    fn load() -> Result<Container> {
+        let container = Container {
+            id: "hoge".into(),
+            bundle: PathBuf::from("."),
+            spec: Spec::default(),
             status: Status::Created,
-            pid: None,
-            bundle: self.bundle.clone(),
-            rootfs: PathBuf::from(self.spec.root.path.clone()),
-            owner: "root".into(), // TODO: getusername
-            annotation: None,
-            created: None, // TODO:
-        })
+        };
+        Ok(container)
+    }
+
+    fn lock(&self, file: &File) -> Result<()> {
+        let fd = file.as_raw_fd();
+        flock(fd, FlockArg::LockExclusive)?;
+        Ok(())
+    }
+
+    fn unlock(&self, file: &File) -> Result<()> {
+        let fd = file.as_raw_fd();
+        flock(fd, FlockArg::Unlock)?;
+        Ok(())
     }
 }
 
@@ -47,9 +98,8 @@ pub mod testutil {
 
     use crate::specutil;
 
-    pub fn init_bundle_dir(container_id: &str) -> Result<PathBuf> {
-        let base = tempfile::tempdir()?.into_path();
-        let bundle = base.join(container_id);
+    pub fn init_bundle_dir() -> Result<PathBuf> {
+        let bundle = tempfile::tempdir()?.into_path();
         fs::create_dir_all(&bundle)?;
         Ok(bundle)
     }
@@ -61,16 +111,24 @@ pub mod testutil {
         Ok(rootfs)
     }
 
-    pub fn init_spec_file(bundle: &PathBuf) -> Result<()> {
+    pub fn init_spec_file(bundle: &PathBuf, rootfs: &PathBuf) -> Result<()> {
         let mut spec = Spec::default();
-        spec.root.path = bundle.clone().to_str().unwrap().to_string();
+        spec.root.path = rootfs.clone().to_str().unwrap().to_string();
 
         specutil::write(&bundle, &spec)?;
         Ok(())
     }
 
-    pub fn cleanup(bundle: &PathBuf) -> Result<()> {
-        fs::remove_dir_all(&bundle)?;
+    pub fn cleanup(bundle: &PathBuf, meta_dir: &PathBuf) -> Result<()> {
+        // bundledir have rootfs, config.json
+        if bundle.exists() {
+            fs::remove_dir_all(&bundle)?;
+        }
+
+        // remove statefile(e.g /run/runt/<container-id>/state.json)
+        if meta_dir.exists() {
+            fs::remove_dir_all(&meta_dir)?;
+        }
         Ok(())
     }
 }
@@ -81,23 +139,26 @@ pub mod test {
 
     use uuid::Uuid;
 
+    use crate::specutil;
+
     #[test]
     fn bundle_should_be_current_dir() {
         let container_id = Uuid::new_v4().to_string();
-        let bundle = testutil::init_bundle_dir(&container_id).unwrap();
+        let bundle = testutil::init_bundle_dir().unwrap();
         let rootfs = testutil::init_rootfs_dir(&bundle).unwrap();
-        let mut spec = Spec::default();
-        spec.root.path = rootfs.to_str().unwrap().into();
+        testutil::init_spec_file(&bundle, &rootfs).unwrap();
+        let spec = specutil::load(&bundle).unwrap();
 
-        let container = Container::new(&container_id, &bundle, spec);
+        let meta_dir = PathBuf::from(DEFAULT_META_ROOT).join(&container_id);
 
-        let state = container.create().unwrap();
+        let mut container = Container::new(&container_id, &bundle, spec);
+        assert_eq!(container.id, container_id);
+        assert_eq!(container.bundle, bundle);
+        assert_eq!(container.status, Status::Creating);
 
-        assert_eq!(state.oci_version, OCI_VERSION.to_string());
-        assert_eq!(state.id, container_id.to_string());
-        assert_eq!(state.bundle, bundle);
-        assert_eq!(state.rootfs, rootfs);
-        assert_eq!(state.status, Status::Created);
-        testutil::cleanup(&bundle).unwrap();
+        assert!(container.create().is_ok());
+
+        assert_eq!(container.status, Status::Created);
+        testutil::cleanup(&bundle, &meta_dir).unwrap();
     }
 }
